+++ CHANGELOG.md
@@ -33,3 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - switched from JsonUtility to Newtonsoft
+
+## [0.1.4] - 2023-06-19
+### Added
+- added full headless define overwrite support
+
+### Changed
+- switched from ScriptableObject Settings to ScriptableSingleton